requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
structlog
python-multipart

pdfminer.six
pypdf
python-docx

numpy
sentence-transformers
faiss-cpu
torch

pytest
pytest-cov

spacy

tenacity
prometheus-fastapi-instrumentator
